# server/doc/swagger/home-life.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Home Life API",
    "version": "1.1",
    "contact": {
      "name": "Dawid",
      "url": "github.com/Remxin/home-life/server",
      "email": "remxin7@gmail.com"
    }
  },
  "tags": [
    {
      "name": "HomeLife"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/create_user": {
      "post": {
        "summary": "Create new user",
        "description": "Use this API to create a new user",
        "operationId": "HomeLife_CreateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbCreateUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbCreateUserRequest"
            }
          }
        ],
        "tags": [
          "HomeLife"
        ]
      }
    },
    "/v1/family": {
      "post": {
        "summary": "Create family group",
        "description": "Use this API to create a family group",
        "operationId": "HomeLife_CreateFamily",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbCreateFamilyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbCreateFamilyRequest"
            }
          }
        ],
        "tags": [
          "HomeLife"
        ]
      }
    },
    "/v1/family/add_user": {
      "post": {
        "summary": "Add user to a family",
        "description": "Use this API to add user to a family",
        "operationId": "HomeLife_AddToFamily",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbAddToFamilyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbAddToFamilyRequest"
            }
          }
        ],
        "tags": [
          "HomeLife"
        ]
      }
    },
    "/v1/login_user": {
      "post": {
        "summary": "Login user",
        "description": "Use this API to login user",
        "operationId": "HomeLife_LoginUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbLoginUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbLoginUserRequest"
            }
          }
        ],
        "tags": [
          "HomeLife"
        ]
      }
    },
    "/v1/verify_user": {
      "get": {
        "summary": "Verify user email",
        "description": "Use this API to verify created user email",
        "operationId": "HomeLife_VerifyUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbVerifyUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "emailId",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "secretCode",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "HomeLife"
        ]
      }
    }
  },
  "definitions": {
    "pbAddToFamilyRequest": {
      "type": "object",
      "properties": {
        "familyId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "canRead": {
          "type": "boolean"
        },
        "canEdit": {
          "type": "boolean"
        },
        "canCreate": {
          "type": "boolean"
        },
        "canModify": {
          "type": "boolean"
        }
      }
    },
    "pbAddToFamilyResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/pbUser"
        },
        "permissions": {
          "$ref": "#/definitions/pbPermissions"
        }
      }
    },
    "pbCreateFamilyRequest": {
      "type": "object",
      "properties": {
        "familyName": {
          "type": "string"
        }
      }
    },
    "pbCreateFamilyResponse": {
      "type": "object",
      "properties": {
        "family": {
          "$ref": "#/definitions/pbFamily"
        },
        "permissionToken": {
          "type": "string"
        }
      }
    },
    "pbCreateUserRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "pbCreateUserResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/pbUser"
        }
      }
    },
    "pbFamily": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "ownerId": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "pbLoginUserRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "pbLoginUserResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/pbUser"
        },
        "sessiondId": {
          "type": "string"
        },
        "accessToken": {
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        },
        "accessTokenExpiresAt": {
          "type": "string",
          "format": "date-time"
        },
        "refreshTokenExpiresAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "pbPermissions": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "canRead": {
          "type": "boolean"
        },
        "canEdit": {
          "type": "boolean"
        },
        "canCreate": {
          "type": "boolean"
        },
        "canModify": {
          "type": "boolean"
        }
      }
    },
    "pbUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "passwordChangedAt": {
          "type": "string",
          "format": "date-time"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "pbVerifyUserResponse": {
      "type": "object",
      "properties": {
        "isVerified": {
          "type": "boolean"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}